rename aHgstfac to aHg_param
